Move all strategies to single file.
