+++ src/libslic3r/GCode.cpp
@@ -14,9 +14,7 @@
 #include "GCode/Thumbnails.hpp"
 #include "GCode/WipeTower.hpp"
 #include "ShortestPath.hpp"
-#include "GCode/ConvexHullPeeling.hpp"
-#include "GCode/Boustrophedon.hpp"
-#include "GCode/BestOfStrategies.hpp"
+#include "GCode/OrderingStrategies.hpp"
 #include "Print.hpp"
 #include "Utils.hpp"
 #include "ClipperUtils.hpp"
