+++ src/libslic3r/GCode/OrderingStrategies.hpp
@@ -1,7 +1,8 @@
-// Shared TSP post-processing utilities for print-object ordering strategies.
+// Print-object ordering strategies and shared TSP post-processing utilities.
+// All strategies, post-processing, and wrappers are consolidated into this single module.
 
-#ifndef slic3r_TSPPostProcessing_hpp_
-#define slic3r_TSPPostProcessing_hpp_
+#ifndef slic3r_OrderingStrategies_hpp_
+#define slic3r_OrderingStrategies_hpp_
 
 #include "../libslic3r.h"
 #include "../Point.hpp"
@@ -11,7 +12,6 @@
 #endif
 
 #include <algorithm>
-#include <unordered_map>
 #include <vector>
 
 namespace Slic3r {
@@ -55,70 +55,7 @@ inline double tsp_max_edge_length(const std::vector<size_t>& path, const Points&
     return mx;
 }
 
-// --- Row grouping + serpentine traversal (shared by Boustrophedon & GridPath) ---
 
-// Group points into rows by Y coordinate and traverse them in alternating direction.
-// Returns the initial path before any post-processing.
-inline std::vector<size_t> row_serpentine_path(const Points& centers, double fraction_of_y_range = 0.1, double min_threshold_um = 1e4)
-{
-    if (centers.empty()) return {};
-
-    size_t n = centers.size();
-
-    // Compute Y range to determine row grouping threshold.
-    double y_min = std::numeric_limits<double>::max();
-    double y_max = -std::numeric_limits<double>::max();
-    for (const auto& p : centers) {
-        double y = static_cast<double>(p.y());
-        if (y < y_min) y_min = y;
-        if (y > y_max) y_max = y;
-    }
-
-    // Row threshold: fraction of Y range, with a minimum floor.
-    double row_threshold = (y_max - y_min) * fraction_of_y_range;
-    if (row_threshold < min_threshold_um) row_threshold = min_threshold_um;
-
-    // Group into rows using hash-map binning by Y coordinate.
-    std::unordered_map<int, std::vector<size_t>> row_map;
-    for (size_t i = 0; i < n; ++i) {
-        int y_key = static_cast<int>(centers[i].y() / row_threshold);
-        row_map[y_key].push_back(i);
-    }
-
-    // Convert to vector of rows with precomputed average Y.
-    struct Row { double avg_y; std::vector<size_t> indices; };
-    std::vector<Row> rows;
-    rows.reserve(row_map.size());
-    for (auto& [key, indices] : row_map) {
-        double sum = 0;
-        for (size_t idx : indices) sum += static_cast<double>(centers[idx].y());
-        rows.push_back({sum / static_cast<double>(indices.size()), std::move(indices)});
-    }
-
-    // Sort rows by average Y.
-    std::sort(rows.begin(), rows.end(), [](const auto& a, const auto& b) {
-        return a.avg_y < b.avg_y;
-    });
-
-    // Sort each row by X coordinate, then traverse alternating direction (serpentine).
-    std::vector<size_t> path;
-    path.reserve(n);
-    bool reverse = false;
-
-    for (auto& [avg_y, row] : rows) {
-        std::sort(row.begin(), row.end(),
-            [&](size_t a, size_t b) { return centers[a].x() < centers[b].x(); });
-
-        if (reverse) {
-            path.insert(path.end(), row.rbegin(), row.rend());
-        } else {
-            path.insert(path.end(), row.begin(), row.end());
-        }
-        reverse = !reverse;
-    }
-
-    return path;
-}
 
 #ifndef SLIC3R_TEST_HARNESS
 
@@ -168,6 +105,33 @@ std::vector<const PrintInstance*> chain_instances_with_core(
 
 #endif // SLIC3R_TEST_HARNESS
 
+// --- Core algorithms (operate on raw Points, return index permutations) ---
+
+// Boustrophedon (snake) ordering: row grouping + serpentine traversal + post-processing.
+std::vector<size_t> boustrophedon_core(const Points& centers);
+
+// Convex hull peeling (onion peeling) with concave hulls.
+std::vector<size_t> convex_hull_peeling_core(const Points& centers);
+
+#ifndef SLIC3R_TEST_HARNESS
+
+// --- Production wrappers ---
+
+// Boustrophedon (snake-like) ordering.
+std::vector<const PrintInstance*> chain_print_object_instances_boustrophedon(const std::vector<const PrintObject*>& print_objects, const Point* start_near);
+std::vector<const PrintInstance*> chain_print_object_instances_boustrophedon(const Print& print);
+
+// Convex hull peeling (onion peeling).
+std::vector<const PrintInstance*> chain_print_object_instances_convex_hull_peeling(const std::vector<const PrintObject*>& print_objects, const Point* start_near);
+std::vector<const PrintInstance*> chain_print_object_instances_convex_hull_peeling(const Print& print);
+
+// Best-of-strategies: run all strategies and return the shortest result.
+// Primary: shortest total path; secondary tiebreaker: smallest max edge.
+std::vector<const PrintInstance*> chain_print_object_instances_best_of(const std::vector<const PrintObject*>& print_objects, const Point* start_near);
+std::vector<const PrintInstance*> chain_print_object_instances_best_of(const Print& print);
+
+#endif // SLIC3R_TEST_HARNESS
+
 } // namespace Slic3r
 
-#endif /* slic3r_TSPPostProcessing_hpp_ */
+#endif /* slic3r_OrderingStrategies_hpp_ */
